Update compiler standard to C++20
Related to #723

This commit changes the compiler, and fixes the problems that
prevented compilation. It does NOT add any of the new C++20
features, it just means the code base can now utilize C++ 20
features.

+++ src/utils/font_prop.cpp
@@ -186,7 +186,7 @@ void FontProperty::Convert(ttlib::cview font)
     if (mstr.size() > font::idx_facename_style)
     {
         auto value = std::atof(std::string(mstr[font::idx_facename_style]).c_str());
-        if (value < wxFONTSTYLE_NORMAL)  // wxFONTSTYLE_NORMAL == 90, so too large to be a point size
+        if (value < static_cast<double>(wxFONTSTYLE_NORMAL))  // wxFONTSTYLE_NORMAL == 90, so too large to be a point size
         {
             m_pointSize = value;
